Extract preview data and K5 font setup to a helper entity.
refs: MBL-15475
affects: Student
release note: none

test plan: Not testable. Helper entity to speed up K5 preview creation.
